parte4

+++ block7-crud-validation/src/main/java/block7crudvalidation/repository/PersonRepository.java
@@ -6,7 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface PersonRepository  extends JpaRepository<Person, Integer>  {
+public interface PersonRepository  extends JpaRepository<Person, Long>  {
     List<Person> findByUsuarioContainingIgnoreCase(String usuario);
+
 
 }
